+++ quacc/recipes/emt/core.py
@@ -5,7 +5,7 @@ import covalent as ct
 from ase import Atoms
 from ase.calculators.emt import EMT
 
-from quacc.schemas.calc import summarize_opt_run, summarize_run
+from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.util.atoms import copy_atoms
 from quacc.util.calc import run_ase_opt, run_calc